requirements/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0
django-cors-headers==3.10.1
django-oauth-toolkit==1.5.0
django-restframework==0.0.1
djangorestframework==3.13.1
djangorestframework-oauth==1.1.0
idna==3.3
jwcrypto==1.0
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.5
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
wrapt==1.13.3